+++ pioreactor/background_jobs/led_control.py
@@ -15,6 +15,7 @@ from contextlib import suppress
 from pioreactor.whoami import get_unit_name, get_latest_experiment_name
 from pioreactor.background_jobs.base import BackgroundJob
 from pioreactor.logging import create_logger
+from pioreactor.background_jobs.utils import AutomationDict
 
 
 class LEDController(BackgroundJob):
@@ -22,28 +23,32 @@ class LEDController(BackgroundJob):
     # this is automagically populated
     automations = {}  # type: ignore
 
-    published_settings = {"led_automation": {"datatype": "string", "settable": True}}
+    published_settings = {
+        "led_automation": {"datatype": "json", "settable": True},
+        "led_automation_key": {"datatype": "string", "settable": False},
+    }
 
-    def __init__(self, led_automation, unit: str, experiment: str, **kwargs):
+    def __init__(self, automation_key, unit: str, experiment: str, **kwargs):
         super(LEDController, self).__init__(
             job_name="led_control", unit=unit, experiment=experiment
         )
 
-        self.led_automation = led_automation
+        self.led_automation = AutomationDict(automation_key=automation_key, **kwargs)
 
         try:
-            automation_class = self.automations[self.led_automation]
+            automation_class = self.automations[self.led_automation["automation_key"]]
         except KeyError:
             raise KeyError(
-                f"Unable to find automation {self.led_automation}. Available automations are {list(self.automations.keys())}"
+                f"Unable to find automation {self.led_automation['automation_key']}. Available automations are {list(self.automations.keys())}"
             )
 
         self.led_automation_job = automation_class(
             unit=self.unit, experiment=self.experiment, **kwargs
         )
+        self.led_automation_key = self.led_automation["automation_key"]
 
     def set_led_automation(self, new_led_automation_json: str):
-        algo_init = json.loads(new_led_automation_json)
+        algo_metadata = AutomationDict(json.loads(new_led_automation_json))
 
         try:
             self.led_automation_job.set_state("disconnected")
@@ -53,10 +58,11 @@ class LEDController(BackgroundJob):
             self.set_led_automation(new_led_automation_json)
 
         try:
-            self.led_automation_job = self.automations[algo_init["led_automation"]](
-                unit=self.unit, experiment=self.experiment, **algo_init
+            self.led_automation_job = self.automations[algo_metadata["automation_key"]](
+                unit=self.unit, experiment=self.experiment, **algo_metadata
             )
-            self.led_automation = algo_init["led_automation"]
+            self.led_automation = algo_metadata
+            self.led_automation_key = self.led_automation["automation_key"]
 
         except Exception as e:
             self.logger.debug(f"Change failed because of {str(e)}", exc_info=True)
@@ -103,7 +109,7 @@ def start_led_control(
     context_settings=dict(ignore_unknown_options=True, allow_extra_args=True),
 )
 @click.option(
-    "--automation",
+    "--automation-key",
     default="silent",
     help="set the automation of the system: silent, etc.",
     show_default=True,
@@ -117,12 +123,12 @@ def start_led_control(
     help="Normally algo will run immediately. Set this flag to wait <duration>min before executing.",
 )
 @click.pass_context
-def click_led_control(ctx, automation, duration, skip_first_run):
+def click_led_control(ctx, automation_key, duration, skip_first_run):
     """
     Start an LED automation
     """
     lc = start_led_control(
-        automation=automation,
+        automation_key=automation_key,
         duration=duration,
         skip_first_run=skip_first_run,
         **{ctx.args[i][2:]: ctx.args[i + 1] for i in range(0, len(ctx.args), 2)},