+++ CHANGELOG.md
@@ -4,6 +4,12 @@
  - fixed errors raised when not able to connect to leader's MQTT
  - improvements to error handling in monitor job.
  - replaced `turn_off_leds_temporarily` with the more useful and more general `change_leds_intensities_temporarily`
+ - UX improvements to the `pump_calibration` action
+ - improving \*-Controller jobs:
+    - `*_automation` is now a dict attribute (json in MQTT)
+    - `*_automation_key` is a new published_setting, with string. This is what is read from the UI.
+    - CLI has a slightly changed API to pick the automation
+  - `pio_jobs_running` renamed to `pio_processes_running`
 
 ### 21.11.1
  - a version cut to test building images